Refactor of components into packages
