+++ app/src/main/java/com/hfad/bitsandpizzas/ui/stores/StoresFragment.java
@@ -1,4 +1,4 @@
-package com.hfad.bitsandpizzas;
+package com.hfad.bitsandpizzas.ui.stores;
 
 
 import android.os.Bundle;
@@ -8,6 +8,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
+
+import com.hfad.bitsandpizzas.R;
 
 
 /**
